+++ tests/TestHomography.cpp
@@ -8,12 +8,12 @@
 namespace wip {
 TEST(HomographyEstimator, test_calculate) {
   wip::HomographyEstimator he;
-  std::vector<cv::Point2f> src;
+  std::vector<cv::Point2d> src;
   src.emplace_back(1, 2);
   src.emplace_back(2, 6);
   src.emplace_back(-3, 0.3);
   src.emplace_back(9, -4);
-  std::vector<cv::Point2f> dst;
+  std::vector<cv::Point2d> dst;
   dst.emplace_back(1.22135867, 0.93899514);
   dst.emplace_back(1.64745532, 1.07120274);
   dst.emplace_back(-189.32242839, -54.92238531);
@@ -34,12 +34,12 @@ TEST(HomographyEstimator, test_calculate) {
 
 TEST(HomographyEstimator, test_evaluate) {
   wip::HomographyEstimator he;
-  std::vector<cv::Point2f> src;
+  std::vector<cv::Point2d> src;
   src.emplace_back(1, 2);
   src.emplace_back(2, 6);
   src.emplace_back(-3, 0.3);
   src.emplace_back(9, -4);
-  std::vector<cv::Point2f> dst;
+  std::vector<cv::Point2d> dst;
   dst.emplace_back(1.22135867, 0.93899514);
   dst.emplace_back(1.64745532, 1.07120274);
   dst.emplace_back(-189.32242839, -54.92238531);
@@ -53,18 +53,18 @@ TEST(HomographyEstimator, test_evaluate) {
 
 TEST(HomographyEstimator, test_getPose) {
   wip::HomographyEstimator he;
-  std::vector<cv::Point2f> src;
+  std::vector<cv::Point2d> src;
   src.emplace_back(1, 2);
   src.emplace_back(2, 6);
   src.emplace_back(-3, 0.3);
   src.emplace_back(9, -4);
-  std::vector<cv::Point2f> dst;
+  std::vector<cv::Point2d> dst;
   dst.emplace_back(1.22135867, 0.93899514);
   dst.emplace_back(1.64745532, 1.07120274);
   dst.emplace_back(-189.32242839, -54.92238531);
   dst.emplace_back(1.89840739, 1.11539421);
 
-  cv::Mat K = (cv::Mat_<float>(3, 3) << 100, 0, 50, 0, 100, 50, 0, 0, 1);
+  cv::Mat K = (cv::Mat_<double>(3, 3) << 100, 0, 50, 0, 100, 50, 0, 0, 1);
 
   // const auto H = he.calculate(src, dst);
   // std::vector<cv::Mat> rotations, translations, normals;
